Move multiplatform JSON assertions test into commonTest.
